@@ -1,4 +1,4 @@
-import {HttpClient} from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 import {Component, OnInit} from '@angular/core';
 import {ResultWithFilterableFields, Selection} from 'projects/ngx-coal/src/lib/components/filter-bar/filter-bar.component';
 import {BehaviorSubject, map, tap} from 'rxjs';
